Change return type of getMaxSize to int, to match the internal variable.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@929389 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/base/src/org/ofbiz/base/util/cache/UtilCache.java
@@ -453,7 +453,7 @@ public class UtilCache<K, V> implements Serializable {
     /** Returns the current maximum number of elements in the cache
      * @return The maximum number of elements in the cache
      */
-    public long getMaxSize() {
+    public int getMaxSize() {
         return maxSize;
     }
